Added eventRegistry class to handle store and handle events
