+++ spilo/channel.py
@@ -1,9 +1,10 @@
 import asyncio
 import pickle
-from typing import Set, Dict, Any, Callable
+from typing import Set, Dict, Any
 
 from .base_client import BaseClient
 from .base_pubsub import BaseAsyncPubSub
+from .event_registry import EventRegistry
 
 
 class Channel:
@@ -15,13 +16,13 @@ class Channel:
 
     _channel_cache = {}
 
-    def __init__(self, channel_name: str, pubsub_manager: BaseAsyncPubSub):
+    def __init__(self, channel_name: str, pubsub_manager: BaseAsyncPubSub, event_registry: EventRegistry):
         self.channel_name = channel_name
         self._clients: Set[BaseClient] = set()
         self._dict_clients: Dict[Any, BaseClient] = {}
         self.pubsub_manager: BaseAsyncPubSub = pubsub_manager
         self._receiver_task: asyncio.Task = asyncio.create_task(self.receiver())
-        self.__events = {}
+        self._event_registry = event_registry
 
     def __len__(self):
         """
@@ -46,18 +47,6 @@ class Channel:
         channel = cls(channel_name, pubsub_manager)
         cls._channel_cache[channel_name] = channel
         return channel
-
-    def on(self, event_name: str = None):
-        """
-        function decorator for handling specific events
-        """
-        def decorator(func: Callable):
-            if event_name:
-                self.__events[event_name] = func
-            else:
-                self.__events[func.__name__] = func
-            return func
-        return decorator
 
     def add_client(self, client: BaseClient) -> None:
         """
