+++ spilo/utils.py
@@ -1,10 +1,9 @@
 def singleton(class_):
     instances = {}
 
-    def get_instanace(*args, **kwargs):
+    def get_instance(*args, **kwargs):
         if class_ not in instances:
             instances[class_] = class_(*args, **kwargs)
         return instances[class_]
 
-    return get_instanace
-
+    return get_instance
